// 06-chat/chat-app/src/enums/Color.ts
export enum Color {
  RED = "#fe4040",
  YELLOW = "#FDC748",
  WHITE = "#fff",
  BONE = "#faf6e7",
  BLACK = "#000",
  GRADIENT = "linear-gradient(90deg, rgba(254,201,30,1) 0%, rgba(254,64,64,1) 100%)",
  ORANGE = "#FE935B",
  PASTELGREEN = "#CCFFCC",
  PASTELVIOLET = "#CCCCFF",
  PASTELSALMON = "#FF7EB9",
  PASTELFUSCIA = "#ff65a3",
  PASTELYELLOW = "#feff9c",
  PASTELBLUE = "#6ed8fa",
}